Use jQuery UI date picker with default theme.
The new date picker is in use on both the story budget and in editorial metadata. We can certainly use a fancier theme (or a lighter weight one without images) in the future, but this is a start.

I added the fewest images necessary to avoid any errors. The actual jQuery UI download contained many more seemingly unneeded images for our purposes.

[Fixes #18437845]