Fix localization of editorial metadata and custom status quick edit fields.
While in general it'd probably be easier for translators to make more accurate translations if we used context in our gettext calls (we currently only do so when registering the editorial metadata taxonomy), we should do so knowingly and think hard about it. Frankly, it'd probably be easier for someone translating to add in context to our translations if/when necessary, since they'll have better context (no pun intended).

[#6456609]