Sorry for the ginormous commit. I intended to keep it large so the code review would be easy but I can see the diff isn't going to be that simple.
There's a lot going on here. Basically, I completely refactored Usergroups to be modular and use the List Table API. Yes, there's a lot of simularity between this, custom statuses, and editorial metadata. It would be nice to consolidate in the future.

It didn't work to use taxonomies with user objects because the user ID sometimes conflicts with post IDs. This would only work if you were using two separate taxonomies. Instead, I've taken the same approach as with Editorial Metadata to save the user IDs as an encoded string in the taxnomy term description. There's a new helper method that supports encoding and decoding the data.

The last bit to do with this story is reenable post subscriptions. As this time, I think we should move the notifications class to be a 'subscriptions' class and have it be a supporting module for all of the notifications/subscriptions in Edit Flow.

[#16865531]