Hide built-in statuses and custom statuses from appearing in the title column on the Manage Posts screen.
We previously were adding custom statuses to this column. That code has been removed. In addition, we now actively hide built-in statuses (like 'Draft') from appearing in this column.

While the status itself is hidden with CSS, there is a remaining dash at the end of the title that we remove with JS after page load. The result is that the dash is visible while the page loads. This oddity can only be fixed with a change to Core to make the dash more easily targetable.

[Fixes #22864139]