Allow the post_name to continue to be influenced by a changing post_title when we're using custom statuses
The problem happens in two parts:

1) Core sets the post_name value if the post_status isn't draft, pending, or auto-draft:
http://core.trac.wordpress.org/browser/tags/3.4.2/wp-includes/post.php#L2646

2) When editing a post, the editable post_name / slug isn't auto-sexily generated if the post_status isn't draft, pending, or auto-draft:
http://core.trac.wordpress.org/browser/tags/3.4.2/wp-admin/includes/post.php#L1023

On the frontend, if the post_name is empty, the editable slug is empty too and looks bad. So we fake it. But, we also use a static var to ensure it's only faked once. If we fake it twice, the post meta box for editing slugs is filled with a fake post_name that ends up getting set as the post_name because of 1.

See #123