Added timepicker to date metadata. Enabled default option to not set the time by default, which I think should offer graceful backwards compatability.
Determine if date or datetime should be displayed.

Added a .date-pick and .date-time-pick class to differentiate between the classic datepicker and the datetimepicker. This is important because the budget uses a datepicker to determine the dates to display and having the ability to choose time could be confusing.

Consolidated styling.

Minor syntax and style fixes.

Make function name more descriptive by changing from show_time to show_date_or_datetime.

Select lists seem to be the preferred method for a datetime picker (plus they appear to be more platfrom/mobile friendly). I wanted to push this commit so the select is enabled by default. In the meantime, it would probably be worthwhile to look into incoporating some of the tools @sbressler and @mbijon mentioned as component types of the datetime picker. That way everything stays in one place and it looks sleek and stylish.

Fixing issues with css and ef_date.js.

Added a .date-pick and .date-time-pick class to differentiate between the classic datepicker and the datetimepicker. This is important because the budget uses a datepicker to determine the dates to display and having the ability to choose time could be confusing.

Consolidated styling.

Select lists seem to be the preferred method for a datetime picker (plus they appear to be more platfrom/mobile friendly). I wanted to push this commit so the select is enabled by default. In the meantime, it would probably be worthwhile to look into incoporating some of the tools @sbressler and @mbijon mentioned as component types of the datetime picker. That way everything stays in one place and it looks sleek and stylish.

Rebased, added @since 0.8 to show_date_or_datetime and made function private.

Fixing rebase issues.

Removed mistakenly committed code.