A follow link for Manage Posts, Calendar, and Story Budget
This adds a "Follow" link to post row actions to Manage Posts, Calendar, and Story budget views. It also removes the unfollowing_users taxonomy, as that hasn't ever been used, and refactors follow_post_users and unfollow_post_users(), as the latter hasn't ever worked.

Lastly, it refactors the row actions for the Story Budget so it's easy to hook on this follow link

See #130