Persist the current value for the `post_date` field when a post is updated.
In `wp_update_post()`, core will reset the `post_date` field if the post currently has a status of `draft`, `pending` or `auto-draft`.
The Edit Flow calendar uses the `post_date` field as the canonical reference for the post's position on the calendar.
To prevent the post from moving to the current date when it's updated, we need to persist the `post_date` value for the post.