Fix post_date_gmt error
With the current code, it sets the post_date_gmt to 0000-00-00 00:00:00 even if the post_status is publish when posts are published programmatically rather than via the admin (for example, via XMLRPC). I don't see any reason to massage the post_date_gmt at all if the status of the post is not a custom post status.