Updating custom status hack to prevent publishing when scheduling
This is a fix for ticket #301. After setting a scheduled datetime for a post and saving it as a custom status and the user makes any changes the autosave fires. This calls the `fix_custom_status_timestamp` function that has a conditional to blank out `$data['post_date_gmt']` if `$_POST['aa']` is empty. `$_POST['aa']` is not passed on autosaves causing WordPress to publish the current post when clicking schedule.