Fix custom status timestamp bugs
When custom statuses are enabled, post_date_gmt needs to be appropriately set or unset. If the status of a post is 'future' or 'publish', we shouldn't interfere with post_date_gmt.

For any other status, we should unset the post_date_gmt if, when the post was saved, the post_date_gmt was unset or equivalent to 0000-00-00 00:00:00.