Build home_url() previews with trailing slash
`home_url()` will only add a trailing slash to a site's home URL
if a path is passed as an argument.

If no path is passed, a page preview could be something like:

http://wp.wsu.dev?page_id=1

And it would work fine in most cases.

In a subdirectory multisite configuration, a page preview could be
something like:

http://wp.wsu.dev/site?page_id=1

This could be processed by WordPress as a lookup for a page named
"site" rather than a site with that URL because no trailing slash
exists.

When adding query args to `home_url()`, we should tell it to use
a slash to avoid confusion.

See `_get_page_link()` for how core builds these in a way where a
slash is always added.