Refactor to a better way to pass variables
Use wp_localize_script() instead of printing script tags in the header. This is better`custom-status.js` relies on the variables anyway and wp_localize_script is going to always print the variables when wp_enqueue_script is called.

Also - this significantly cleans up global scope and allows i18n