Changed the structure of the list from the Story Budget page
The columns on this page were created through PHP file by wrapping each column into a different div. This way, the width couldn't be changed dinamically depending on the number of columns was set on the Screen Options side. Now, I changed this part to be a big list and change the width of the items through CSS by simply modifying the class of the big wrapper.