Fixes the dash between the post title and status label
When the custom statuses are used, there's an option to display a column with the statuses on Posts section (on dashboard). Right now, the dashes between the post title and status label (that are hidden) are still there. The class for post titles was wrong and I replaced the regex with a substr function. This fixes #395.