Added PHP 5.2 and PHP 5.3 in the tests
These two versions of PHP (5.2 and 5.3) won't be supported by the theme anymore but, for now, they are added as "allowed to fail". Another change that I've made is the check for the PHPUnit test version. Right now, the 5.7.* version is installed if the PHP is 7.x and PHPUnit 4.8.* is installed otherwise.