Added a test for multisite preview page
This change is meant to test the PR #357. Basically, what I'm trying to do is check if the current installation is multisite, create a new blog, move to it, create a new page and check its preview link. The problem is that the preview link seems to have a broken structure once I create the new blog.