Refactor for the fix_post_row_actions function
There code from lines 1649 to 1664 are the lines of the get_preview_link function. That's why I removed all this code and replaced it with the call of the get_preview_link. Another change made in this PR is making this get_preview_link method static. The purpose of this function is to change the actions of a post. Also, making it static will let us use it outside this clas at any moment.

e Please enter the commit message for your changes. Lines starting