Fix status list to show custom post statuses
Fixes #445

Previously, the `get_post_status_friendly_name` function was retreiving the WordPress built-in Post Statuses and the Custom Statuses set within the plugin interface.  If you happen to register a post status another way, it would not appear in the Status Column of your Post Type.

This fix simplifies the logic to just get all the post stati, built-in, registered through the plugin interface, and registered on their own.  It returns the label that was set for each status.