Fix bug preventing previewing posts authored by other users
Bug description:

The author of a post can preview changes to their own published posts,
but other users see the content of the published post when previewing.

This is due to checking for the autosaves of `$query_args->post_author`
instead of the autosaves of `get_current_user_id()`.

Steps to reproduce:

1. Log in as user A
2. Create a new post, press "Publish"
3. Log in as user B
4. Visit the edit page for the post that was just created
5. Add to the post's content (without pressing "Update")
6. Click "Preview"

Expected behaviour:

The content added by user B appears in the preview.

Actual behaviour:

The content added by user B doesn't appear in the preview of the post.